Update README
Update the readme to include steps to install dependencies for the
express backend. Reorganize sections to make some specifications for the
database seperate from other selenium steps. This helps to make it more
clear what each component is used for and how they need to be set up to
interact.